# swagger.json
{
...
  "paths": {
    "/users": {
      "get": {
        "summary": "Get all users",
        "description": "Get all users",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "Returns all data for all users",
            "x-oad-type": "response"
          }
        }
      },
      "post": {
        "summary": "New User",
        "description": "Sign up and make a user",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "Sign up",
            "x-oad-type": "response"
          }
        },
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "description": "name of user",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "password for user",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          }
        ]
      },
      "x-oad-type": "operation"
    },
...
    "/login": {
      "post": {
        "summary": "allows user to login",
        "description": "set request_key for user so user can do actions",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "returns key that user can send along with requests to allow actions",
            "x-oad-type": "response"
          }
        },
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "description": "username of user logging in",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "password used to authenticate user",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          }
        ]
      },
      "delete": {
        "summary": "logout",
        "description": "removes request_key data from user. dissallowing requests from that user",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "logged out",
            "x-oad-type": "response"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "description": "id of user signing out",
            "required": true,
            "type": "integer",
            "format": "int32",
            "x-oad-type": "parameter"
          },
          {
            "name": "request_key",
            "in": "header",
            "description": "passed to authenticate user",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          }
        ]
      },
      "x-oad-type": "operation"
    },
    "/leagues": {
      "get": {
        "summary": "Get all leagues",
        "description": "Get all leagues",
        "tags": [
          "Organizers"
        ],
        "responses": {
          "200": {
            "description": "Returns all data for all leagues",
            "x-oad-type": "response"
          }
        }
      },
      "post": {
        "summary": "New League",
        "description": "Create a new organizer",
        "tags": [
          "Organizers"
        ],
        "responses": {
          "200": {
            "description": "New organizer",
            "x-oad-type": "response"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "description": "name of league",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "owner_id",
            "in": "formData",
            "description": "owner user",
            "required": true,
            "type": "integer",
            "x-oad-type": "parameter"
          },
          {
            "name": "organizer_key",
            "in": "formData",
            "description": "key used for edits. Give it to admins of organizer",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "entry_fee",
            "in": "formData",
            "description": "money in cents",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "description",
            "in": "formData",
            "description": "description of league",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "request_key",
            "in": "header",
            "description": "key passed to allow actions",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          }
        ]
      },
      "x-oad-type": "operation"
    },
...
    "/teams": {
      "get": {
        "summary": "get all teams",
        "description": "get all teams",
        "tags": [
          "Teams"
        ],
        "responses": {
          "200": {
            "description": "get all teams data",
            "x-oad-type": "response"
          }
        }
      },
      "post": {
        "summary": "New team",
        "description": "makes a new team. only works if you are off a team",
        "tags": [
          "Teams"
        ],
        "responses": {
          "200": {
            "description": "all team data for all teams",
            "x-oad-type": "response"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "description": "name of team",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "team_key",
            "in": "formData",
            "description": "join key so others can join your team. set this to a code you can pass to your friends",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "description",
            "in": "formData",
            "description": "description of team",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "user_id",
            "in": "formData",
            "description": "id of owner",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "request_key",
            "in": "header",
            "description": "key authenticating user",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          }
        ]
      },
      "x-oad-type": "operation"
    },
...
    "/tournaments": {
      "get": {
        "summary": "get all tournaments",
        "description": "gets all tournament data for all tournaments",
        "tags": [
          "Tournaments"
        ],
        "responses": {
          "200": {
            "description": "all data for tournaments",
            "x-oad-type": "response"
          }
        }
      },
      "post": {
        "summary": "create new tournament",
        "description": "creats a new tournament",
        "tags": [
          "Tournaments"
        ],
        "responses": {
          "200": {
            "description": "made a new tournament",
            "x-oad-type": "response"
          }
        },
        "parameters": [
          {
            "name": "organizer_id",
            "in": "formData",
            "description": "link to organizer",
            "required": true,
            "type": "integer",
            "format": "int32",
            "x-oad-type": "parameter"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "the name of the tournament",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "type",
            "in": "formData",
            "description": "type of tournament.",
            "required": true,
            "type": "integer",
            "format": "int32",
            "x-oad-type": "parameter"
          },
          {
            "name": "size",
            "in": "formData",
            "description": "max number of teams",
            "required": true,
            "type": "integer",
            "format": "int32",
            "x-oad-type": "parameter"
          },
          {
            "name": "start_date",
            "in": "formData",
            "description": "start date of tournament",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "entry_fee",
            "in": "formData",
            "description": "price to enter tournament",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "description",
            "in": "formData",
            "description": "description of tournament",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "request_key",
            "in": "header",
            "description": "authentication in case of owner",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "organizer_key",
            "in": "header",
            "description": "code passed that allows for organizer admin privleges",
            "required": false,
            "type": "string",
            "x-oad-type": "parameter"
          }
        ]
      },
      "x-oad-type": "operation"
    },
...
    "/matches/{id}": {
      "get": {
        "summary": "get one match",
        "description": "get all match data for one match",
        "tags": [
          "Matches"
        ],
        "responses": {
          "200": {
            "description": "gets all data for one match",
            "x-oad-type": "response"
          }
        }
      },
      "put": {
        "summary": "edit match",
        "description": "edit match",
        "tags": [
          "Matches"
        ],
        "responses": {
          "200": {
            "description": "successfully edited match",
            "x-oad-type": "response"
          }
        },
        "parameters": [
          {
            "name": "tournament_id",
            "in": "formData",
            "description": "tournament match is apart of",
            "required": false,
            "type": "integer",
            "format": "int32",
            "x-oad-type": "parameter"
          },
          {
            "name": "home_id",
            "in": "formData",
            "description": "home team being added to match",
            "required": false,
            "type": "integer",
            "format": "int32",
            "x-oad-type": "parameter"
          },
          {
            "name": "away_id",
            "in": "formData",
            "description": "away team being added to match",
            "required": false,
            "type": "integer",
            "format": "int32",
            "x-oad-type": "parameter"
          },
          {
            "name": "home_score",
            "in": "formData",
            "description": "home score",
            "required": false,
            "type": "integer",
            "format": "int32",
            "x-oad-type": "parameter"
          },
          {
            "name": "away_score",
            "in": "formData",
            "description": "home score",
            "required": false,
            "type": "integer",
            "format": "int32",
            "x-oad-type": "parameter"
          },
          {
            "name": "start_date",
            "in": "formData",
            "description": "start date for match",
            "required": false,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "end_date",
            "in": "formData",
            "description": "start date for match",
            "required": false,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "request_key",
            "in": "header",
            "description": "key to authenticate user",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "organizer_key",
            "in": "header",
            "description": "key used to authenticate organizer admin",
            "required": false,
            "type": "string",
            "x-oad-type": "parameter"
          }
        ]
      },
      "delete": {
        "summary": "delete match",
        "description": "delete match",
        "tags": [
          "Matches"
        ],
        "responses": {
          "200": {
            "description": "match deleted",
            "x-oad-type": "response"
          }
        },
        "parameters": [
          {
            "name": "request_key",
            "in": "header",
            "description": "key to authenticate user",
            "required": true,
            "type": "string",
            "x-oad-type": "parameter"
          },
          {
            "name": "organizer_key",
            "in": "header",
            "description": "key to authenticate organizer admin",
            "required": false,
            "type": "string",
            "x-oad-type": "parameter"
          }
        ]
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "id of match",
          "required": true,
          "type": "integer",
          "format": "int32",
          "x-oad-type": "parameter"
        }
      ],
      "x-oad-type": "operation"
    }
  },
  "info": {
    "title": "Electeams API",
    "version": "1.0.0",
    "description": "Feeder API for electeams",
    "contact": {
      "name": "Caleb",
      "email": "calebgmahala@gmail.com"
    }
  },
  "host": "electron-schedule.herokuapp.com",
  "basePath": "/",
  "swagger": "2.0"
}